Update Search files
