+++ src/main/java/com/exchange/buysell/entity/Data.java
@@ -1,29 +1,6 @@
 package com.exchange.buysell.entity;
 
-import java.util.List;
-
-
-class MyJSONObject {
-    private List<Data> dataList;
-
-    public MyJSONObject() { }
-
-    public MyJSONObject(List<Data> dataList) {
-        this.dataList = dataList;
-    }
-
-    public List<Data> getDataList() {
-        return dataList;
-    }
-
-    public void setDataList(List<Data> dataList) {
-        this.dataList = dataList;
-    }
-}
-
-
 public class Data {
-
     private String symbol;
 
     public Data() { }
@@ -40,3 +17,4 @@ public class Data {
         this.symbol = symbol;
     }
 }
+
