+++ src/main/java/com/exchange/buysell/services/SearchServiceImp.java
@@ -1,38 +1,27 @@
 package com.exchange.buysell.services;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
+import com.exchange.buysell.entity.Data;
+import com.exchange.buysell.entity.MyJSONObject;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import org.springframework.web.client.RestTemplate;
 
+import java.util.List;
+
 
 @Service
 public class SearchServiceImp implements SearchInterface {
-    private  static  final Logger log= LoggerFactory.getLogger(SearchServiceImp.class);
+    private final String URL = "http://api.worldtradingdata.com/api/v1/stock?symbol=AB&api_token=cYDw6OCIFS7okSkLnHj4oR6kP5H3pbQJNBQ9RIFUKZ6MOZYQxAHHhaBWsTfx";
 
+    public List<Data> StockBySymbol(String symbol){
+        RestTemplate restTemplate = new RestTemplate();
+        MyJSONObject myJSONObject = restTemplate.getForObject(URL, MyJSONObject.class);
 
+        System.out.println(myJSONObject);
 
-//    @Override
-//    public List<data> getStocks(@RequestParam String stockname) {
-//      //  String stockSymbol=stockname.toLowerCase();
-//        log.info("In get Stocks");
-//        RestTemplate rt =new RestTemplate();
-//      data data = rt.getForObject("http://api.worldtradingdata.com/api/v1/stock?symbol="+stockname+"&api_token=cYDw6OCIFS7okSkLnHj4oR6kP5H3pbQJNBQ9RIFUKZ6MOZYQxAHHhaBWsTfx", data.class);
-//        System.out.println(data.getPrice());
-//
-//        List<data> mylist = new ArrayList<>();
-//            mylist.add(data);
-//
-//       // mylist.stream().filter(s->s.getStockname().equals(stockname));
-//
-//        return mylist;
-//    }
-    public ResponseEntity<String> getStocks(String stockname){
-        RestTemplate rt =new RestTemplate();
-        ResponseEntity<String> st = rt.getForEntity("http://api.worldtradingdata.com/api/v1/stock?symbol={stockname}&api_token=cYDw6OCIFS7okSkLnHj4oR6kP5H3pbQJNBQ9RIFUKZ6MOZYQxAHHhaBWsTfx", String.class,stockname);
-    return  st;
+        return myJSONObject.getDataList();
     }
+
 }
